added if condition for empty dataframes
added if condition for empty dataframes s.t. bot does not crash. If there is a pair for which no price is available or the pair is not available, the pool-id will be added to pending_pools and hence this pool will not be considered in the next iteration anymore. Also as evan implemented the record function, all pending pools are also recorded.